// packages/core/error/types.tsx
export interface ErrorTypesInterface {
  CART_NO_ASSIGNED_UNIT: string;
  CART_ISSUES_CREATNG_ORDER: string;
  SELLER_INFO_NOT_AVAILABLE: string;
  SELLER_UNIT_NOT_FOUND: string;
  USER_INFO_NOT_AVAILABLE: string;
  USER_ISSUES_CREATING_ADDRESS: string;
  USER_ISSUES_SETTING_DEFAULT_ADDRESS: string;
  CATALOG_NOT_AVAILABLE: string;
  AUTHOS_ERROR_CREATING_SESSION: string;
  AUTHOS_ERROR_UPDATING_SESSION: string;
  AUTHOS_ERROR_DECODE_JWT: string;
  AUTHOS_ERROR_TABLE_NOT_FOUND: string;
  AUTHOS_ERROR_INVALID_SESSION: string;
  AUTHOS_ERROR_EMAIL_ALREADY_REGISTERED: string;
  AUTHOS_ERROR_CREATING_ECOMM_USER: string;
  AUTHOS_ERROR_ELEMENT_NOT_FOUND: string;
  AUTHOS_ERROR_WRONG_PASSWORD: string;
  AUTHOS_ERROR_CREATING_RESTORE: string;
  AUTHOS_ERROR_INVALID_RESTORE: string;
  AUTHOS_ERROR_USER_ALREADY_LOGGED: string;
  AUTHOS_ERROR: string;
  SUPPLIER_PRODUCT_ID_IS_EMPTY: string;
}

export const ErrorCodes: Record<string, string> = {
  11000: "AUTHOS_ERROR",
  11001: "AUTHOS_ERROR_CREATING_SESSION",
  11002: "AUTHOS_ERROR_UPDATING_SESSION",
  11003: "AUTHOS_ERROR_DECODE_JWT",
  11004: "AUTHOS_ERROR_TABLE_NOT_FOUND",
  11005: "AUTHOS_ERROR_INVALID_SESSION",
  11006: "AUTHOS_ERROR_EMAIL_ALREADY_REGISTERED",
  11007: "AUTHOS_ERROR_CREATING_ECOMM_USER",
  11008: "AUTHOS_ERROR_ELEMENT_NOT_FOUND",
  11009: "AUTHOS_ERROR_WRONG_PASSWORD",
  11010: "AUTHOS_ERROR_CREATING_RESTORE",
  11011: "AUTHOS_ERROR_INVALID_RESTORE",
  11012: "AUTHOS_ERROR_USER_ALREADY_LOGGED",
};